Extract `S3HostedJdk`
