+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/flow/start/HookedJiraStartIT.kt
@@ -10,18 +10,15 @@ import com.atlassian.performance.tools.infrastructure.api.jira.flow.install.Defa
 import com.atlassian.performance.tools.infrastructure.api.jira.flow.install.HookedJiraInstallation
 import com.atlassian.performance.tools.infrastructure.api.jira.flow.install.ParallelInstallation
 import com.atlassian.performance.tools.infrastructure.api.jira.flow.server.StartedJira
-import com.atlassian.performance.tools.infrastructure.api.jvm.Jstat
-import com.atlassian.performance.tools.infrastructure.api.jvm.VersionedJavaDevelopmentKit
+import com.atlassian.performance.tools.infrastructure.api.jvm.S3HostedJdk
 import com.atlassian.performance.tools.infrastructure.toSsh
 import com.atlassian.performance.tools.infrastructure.ubuntu.EarlyUbuntuSysstat
 import com.atlassian.performance.tools.jvmtasks.api.Backoff
-import com.atlassian.performance.tools.jvmtasks.api.IdempotentAction
 import com.atlassian.performance.tools.ssh.api.SshConnection
 import com.atlassian.performance.tools.sshubuntu.api.SshUbuntuContainer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.io.File
-import java.net.URI
 import java.nio.file.Files
 import java.time.Duration
 import java.util.function.Consumer
@@ -142,47 +139,8 @@ class HookedJiraStartIT {
 private class FailingHook : PostStartHook {
     override fun run(ssh: SshConnection, jira: StartedJira, flow: PostStartFlow) {
         throw Exception("Expected failure")
-
     }
 }
-
-/**
- * Harvested from https://stash.atlassian.com/projects/JIRASERVER/repos/jira-performance-tests/pull-requests/630
- */
-class S3HostedJdk : VersionedJavaDevelopmentKit {
-    private val jdkVersion = "1.8.0"
-    private val jdkUpdate = 131
-    private val jdkArchive = "jdk${jdkVersion}_$jdkUpdate-linux-x64.tar.gz"
-    private val jdkUrl = URI.create("https://s3.amazonaws.com/packages_java/$jdkArchive")
-    private val jdkBin = "~/jdk${jdkVersion}_$jdkUpdate/jre/bin/"
-    private val bin = "~/jdk${jdkVersion}_$jdkUpdate/bin/"
-    override val jstatMonitoring = Jstat(bin)
-
-    override fun getMajorVersion() = 8
-
-    override fun install(connection: SshConnection) {
-        download(connection)
-        connection.execute("tar -xzf $jdkArchive")
-        connection.execute("echo '${use()}' >> ~/.bashrc")
-    }
-
-    private fun download(connection: SshConnection) {
-        IdempotentAction("download JDK") {
-            connection.execute(
-                cmd = "curl -s -L -O -k $jdkUrl",
-                timeout = Duration.ofMinutes(4)
-            )
-        }.retry(
-            maxAttempts = 3,
-            backoff = StaticBackoff(Duration.ofSeconds(4))
-        )
-    }
-
-    override fun use(): String = "export PATH=$jdkBin:$bin:${'$'}PATH"
-
-    override fun command(options: String) = "${jdkBin}java $options"
-}
-
 
 class StaticBackoff(
     private val backoff: Duration
